@@ -10,8 +10,8 @@
     "demo": "node examples/index.js",
     "deps": "npm-check -u",
     "dev": "nodemon --debug dev/index.js",
-    "lint": "eslint --ext=.js src",
-    "lint:fix": "eslint --fix --ext=.js src",
+    "lint": "eslint --ext=.js src test",
+    "lint:fix": "eslint --fix --ext=.js src test",
     "perf": "nodemon --allow-natives-syntax benchmark/perf-runner.js",
     "pperf": "node --inspect --expose-gc benchmark/perf-runner.js",
     "proto": "pbjs -t static-module -w commonjs -o src/serializers/proto/packets.proto.js src/serializers/proto/packets.proto",
@@ -42,8 +42,10 @@
     "benchmarkify": "2.1.0",
     "bunyan": "1.8.12",
     "coveralls": "2.13.1",
-    "eslint": "4.4.1",
+    "eslint": "4.5.0",
+    "eslint-plugin-node": "^5.1.1",
     "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-security": "^1.4.0",
     "fakerator": "0.3.0",
     "ioredis": "3.1.4",
     "jest": "20.0.4",
