@@ -7,7 +7,7 @@ let ServiceBroker = require("../src/service-broker");
 // Create broker
 let broker = new ServiceBroker({
 	nodeID: "hot-" + process.pid,
-	transporter: "amqp://192.168.0.181:5672",
+	transporter: "amqp://192.168.51.29:5672",
 	logger: console,
 	//logLevel: "debug",
 	hotReload: true
